+++ src/services/user/index.ts
@@ -1,7 +1,7 @@
 import { IUser } from "./types";
 import { withService } from "@/services/with-service";
 
-export const {getUsers, getUserWithoutTheseUsers}: IUser = await withService(
+export const {getUsers, getUserWithoutTheseUsers, login, register}: IUser = await withService(
     () => import("./impl/mock"),
     () => import("./impl/real")
 )
